# Simplify campaign factory