--- a/spec/components/app_activity_log_component_spec.rb
+++ b/spec/components/app_activity_log_component_spec.rb
@@ -87,7 +87,7 @@ describe AppActivityLogComponent, type: :component do
   end
 
   include_examples "card",
-                   title: "Vaccinated with Cervarix (HPV)",
+                   title: "Vaccinated with Gardasil 9 (HPV)",
                    date: "31 May 2024 at 12:00pm",
                    notes: "Some notes",
                    by: "Nurse Joy"