This makes a number of changes to the campaign factory to simplify the default set up (reducing the number of associations) and updates the tests to only use a specific trait when required. Locally this has improved the performance of the tests by about 15 seconds:

## Before

```
Executed in   71.39 secs    fish           external
   usr time   47.79 secs    0.09 millis   47.79 secs
   sys time    3.35 secs    2.15 millis    3.35 secs
```

## After

```
Executed in   56.78 secs    fish           external
   usr time   38.35 secs    0.09 millis   38.35 secs
   sys time    2.64 secs    2.07 millis    2.63 secs

```